requirements.txt
bcrypt==3.1.7
boto3==1.11.11
botocore==1.14.11
bunch==1.0.1
cffi==1.13.2
cryptography==2.8
docutils==0.15.2
future==0.18.2
iso8601==0.1.12
jmespath==0.9.4
munch==2.5.0
paho-mqtt==1.5.0
paramiko==2.7.1
pigpio==1.45
pycparser==2.19
pygame==1.9.6
PyNaCl==1.3.0
pyserial==3.4
python-dateutil==2.8.1
PyYAML==5.3
RPi.GPIO==0.7.0
s3transfer==0.3.2
scp==0.13.2
serial==0.0.97
six==1.14.0
smbus==1.1.post2
urllib3==1.25.8
VL53L0X-rasp-python==1.0.2
